slot-response-production_handoff/requirements.txt
joblib
lightgbm
numpy
pandas
pyarrow
pyodps
PyYAML
scikit-learn
